requirements.txt
openai==1.68.2
python-dotenv==1.0.1
streamlit==1.43.2
